// savon-app/src/app/models/RecetteDTO.ts
import { LigneIngredient } from "./LigneIngredient";

export class RecetteDTO {
  id: number | null = null;
  titre: string = '';
  description: string = '';
  surgraissage: number = 0;
  avecSoude: boolean = false;
  concentrationAlcalin: number = 0;
  qteAlcalin: number = 0;
  ligneIngredients: LigneIngredient[] = [];
  apportEnEau: number = 0;
  resultats: any = {};


}